+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Introduced possibility to schedule an auto-deployment of the Build [#1311](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1311)
 - Allow downloading build logs for Builds in Unknown status [#1309](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1309)
 - Aligned Build removal confirmation message with SAP's one [#1310](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1310)
+- Enable `mode` & `strategy` only if `Deploy` is selected in the Build dialog [#1312](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1312)
 
 ## [2024.3.5]